test/backend_test/forumcore/user/views.py
from django.shortcuts import render
from rest_framework import viewsets
from rest_framework.generics import get_object_or_404
from rest_framework.response import Response

from .serializers import UserSerializer
from .models import UserClass

# Create your views here.
class UserViewSet(viewsets.ViewSet):
    def list(self, request):
        queryset = UserClass.objects.all()
        serializer = UserSerializer(queryset, many=True)
        return Response(serializer.data)

    def retrieve(self, request, pk=None):
        queryset = UserClass.objects.all()
        user = get_object_or_404(queryset, pk=pk)
        serializer = UserSerializer(user)
        return Response(serializer.data)
