test/backend_test/requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
greenlet==3.2.3
iniconfig==2.1.0
msgpack==1.1.1
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.9.0
pynvim==0.5.2
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
pytz==2025.2
rest-pandas==1.1.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
